# web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Swipe Shopify"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-access-token"
            }
        }
    },
    "security": [
        {
            "apiKeyAuth": []
        }
    ],
    "paths": {
...
        "/v1/datafix/webhook-list/{shop}": {
            "get": {
                "tags": ["Datafix"],
                "summary": "Datafix Get API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shop",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "No Records found"
                    }
                }
            }
        },
        "/v1/datafix/webhook-register/{shop}": {
            "get": {
                "tags": ["Datafix"],
                "summary": "Datafix Get API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shop",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "No Records found"
                    }
                }
            }
        },
        "/v1/datafix/usage-record-create/{shop}/{order}": {
            "get": {
                "tags": ["Datafix"],
                "summary": "Datafix Get API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shop",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "order",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],

                "responses": {
                    "200": {
                        "description": "Data found."
                    },
                    "400": {
                        "description": "No Records found"
                    }
                }
            }
        },
        "/v1/datafix/order-create": {
            "post": {
                "tags": ["Datafix"],
                "summary": "Find-Datafix API",
                "requestBody": {
                    "description": "Find Datafix Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Find-Datafix"
                            }
                        }
                    }
                },
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Datafix Found successfully"
                    },
                    "400": {
                        "description": "Datafix Not Found"
                    }
                }
            }
        },
        "/v1/datafix/remove-merchant-data/{shop}": {
            "delete": {
                "tags": ["datafix"],
                "summary": "Delete datafix API",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "shop",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant successfully Deleted"
                    },
                    "400": {
                        "description": "You could not be Deleted Logged-in Merchant"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Find-Datafix": {
            "type": "object",
            "properties": {
                "shop": {
                    "type": "String",
                    "example": "323232"
                },
                "order": {
                    "type": "string",
                    "example": "aaaaaaaa"
                }
            }
        },
...
    }
}